chore(refactor): move tg bot exmaple to root folder
